requirements.txt
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
codecov==2.1.11
coverage==5.5
icdiff==1.9.1
idna==2.10
importlib-metadata==4.0.1
iniconfig==1.1.1
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pluggy==0.13.1
pprintpp==0.4.0
psycopg2-binary==2.8.4
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-codecov==0.3.0
pytest-cov==2.12.0
pytest-icdiff==0.5
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rethinkdb==2.4.8
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
zipp==3.4.1
